pyproject.toml
[project]
name = "fastapi-course"
version = "0.1.0"
description = "Python Rest API Course using FastAPI"
authors = [
    {name = "Ajan Shrestha", email = "ajan.shresh@gmail.com"}
]
package-mode = false
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "ruff (>=0.9.3,<0.10.0)",
    "fastapi[standard] (>=0.115.7,<0.116.0)",
    "psycopg[binary] (>=3.2.4,<4.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"